+++ src/polymer/expression-scanner.ts
@@ -38,6 +38,15 @@ const isDataBindingTemplate = p.AND(
 
 export interface Template extends parse5.ASTNode { content: parse5.ASTNode; }
 
+/**
+ * Given a node, return all databinding templates inside it.
+ *
+ * A template is "databinding" if polymer databinding expressions are expected
+ * to be evaluated inside. e.g. <template is='dom-if'> or <dom-module><template>
+ *
+ * Results include both direct and nested templates (e.g. dom-if inside
+ * dom-module).
+ */
 export function getAllDataBindingTemplates(node: parse5.ASTNode) {
   return dom5.queryAll(
       node,
@@ -49,7 +58,7 @@ export function getAllDataBindingTemplates(node: parse5.ASTNode) {
 /**
  * A databinding expression.
  */
-export class ScannedDatabindingExpression {
+export class DatabindingExpression {
   /**
    * If databinding into an attribute this is the element whose attribute is
    * assigned to. If databinding into a text node, this is that text node.
@@ -78,6 +87,8 @@ export class ScannedDatabindingExpression {
    */
   readonly eventName: string|undefined;
 
+  private readonly locationOffset: LocationOffset;
+
   constructor(
       astNode: parse5.ASTNode, attribute: parse5.ASTAttribute|undefined,
       sourceRange: SourceRange, direction: '{'|'[', expressionText: string,
@@ -92,24 +103,57 @@ export class ScannedDatabindingExpression {
     this.expressionText = expressionText;
     this.expressionAst = ast;
     this.eventName = eventName;
+    this.locationOffset = {
+      line: sourceRange.start.line,
+      col: sourceRange.start.column
+    };
+  }
+
+  /**
+   * Given an estree node in this databinding expression, give its source range.
+   */
+  sourceRangeForNode(node: estree.Node) {
+    if (!node || !node.loc) {
+      return;
+    }
+    const databindingRelativeSourceRange = {
+      file: this.sourceRange.file,
+      // Note: estree uses 1-indexed lines, but SourceRange uses 0 indexed.
+      start: {line: (node.loc.start.line - 1), column: node.loc.start.column},
+      end: {line: (node.loc.end.line - 1), column: node.loc.end.column}
+    };
+    return correctSourceRange(
+        databindingRelativeSourceRange, this.locationOffset);
   }
 }
 
 /**
  * Find and parse Polymer databinding expressions in HTML.
  */
-export function scanForExpressions(document: ParsedHtmlDocument) {
-  const results: ScannedDatabindingExpression[] = [];
+export function scanDocumentForExpressions(document: ParsedHtmlDocument) {
+  return extractDataBindingsFromTemplates(
+      document, getAllDataBindingTemplates(document.ast));
+}
+
+export function scanDatabindingTemplateForExpressions(
+    document: ParsedHtmlDocument, template: Template) {
+  return extractDataBindingsFromTemplates(
+      document,
+      [template].concat(getAllDataBindingTemplates(template.content)));
+}
+
+function extractDataBindingsFromTemplates(
+    document: ParsedHtmlDocument, templates: Iterable<Template>) {
+  const results: DatabindingExpression[] = [];
   const warnings: Warning[] = [];
-  const dataBindingTemplates = getAllDataBindingTemplates(document.ast);
-  for (const template of dataBindingTemplates) {
+  for (const template of templates) {
     dom5.nodeWalkAll(template.content, (node) => {
       if (dom5.isTextNode(node) && node.value) {
-        _extractDataBindingsFromTextNode(document, node, results, warnings);
+        extractDataBindingsFromTextNode(document, node, results, warnings);
       }
       if (node.attrs) {
         for (const attr of node.attrs) {
-          _extractDataBindingsFromAttr(document, node, attr, results, warnings);
+          extractDataBindingsFromAttr(document, node, attr, results, warnings);
         }
       }
       return false;
@@ -118,10 +162,10 @@ export function scanForExpressions(document: ParsedHtmlDocument) {
   return {expressions: results, warnings};
 }
 
-function _extractDataBindingsFromTextNode(
+function extractDataBindingsFromTextNode(
     document: ParsedHtmlDocument,
     node: parse5.ASTNode,
-    results: ScannedDatabindingExpression[],
+    results: DatabindingExpression[],
     warnings: Warning[]) {
   const text = node.value || '';
   const dataBindings = findDatabindingInString(text);
@@ -129,7 +173,10 @@ function _extractDataBindingsFromTextNode(
     return;
   }
   const newlineIndexes = findNewlineIndexes(text);
-  const nodeSourceRange = document.sourceRangeForNode(node)!;
+  const nodeSourceRange = document.sourceRangeForNode(node);
+  if (!nodeSourceRange) {
+    return;
+  }
   // We have indexes into the text node, we'll want to correct that so that
   // it's a range relative to the start of the document.
   const startOfTextNodeOffset: LocationOffset = {
@@ -145,20 +192,16 @@ function _extractDataBindingsFromTextNode(
     const sourceRange =
         correctSourceRange(sourceRangeWithinTextNode, startOfTextNodeOffset)!;
 
-    const dataBindingStartOffset: LocationOffset = {
-      line: sourceRange.start.line,
-      col: sourceRange.start.column
-    };
-    const parseResult = parseJs(
-        dataBinding.expressionText,
-        nodeSourceRange.file,
-        dataBindingStartOffset,
-        'polymer-expression-parse-error');
+    const parseResult =
+        parseExpression(dataBinding.expressionText, sourceRange);
 
+    if (!parseResult) {
+      continue;
+    }
     if (parseResult.type === 'failure') {
       warnings.push(parseResult.warning);
     } else {
-      results.push(new ScannedDatabindingExpression(
+      results.push(new DatabindingExpression(
           node,
           undefined,
           sourceRange,
@@ -173,18 +216,21 @@ function _extractDataBindingsFromTextNode(
   }
 }
 
-function _extractDataBindingsFromAttr(
+function extractDataBindingsFromAttr(
     document: ParsedHtmlDocument,
     node: parse5.ASTNode,
     attr: parse5.ASTAttribute,
-    results: ScannedDatabindingExpression[],
+    results: DatabindingExpression[],
     warnings: Warning[]) {
   if (!attr.value) {
     return;
   }
   const dataBindings = findDatabindingInString(attr.value);
   const attributeValueRange =
-      document.sourceRangeForAttributeValue(node, attr.name, true)!;
+      document.sourceRangeForAttributeValue(node, attr.name, true);
+  if (!attributeValueRange) {
+    return;
+  }
   const attributeOffset: LocationOffset = {
     line: attributeValueRange.start.line,
     col: attributeValueRange.start.column
@@ -211,19 +257,14 @@ function _extractDataBindingsFromAttr(
         newlineIndexes);
     const sourceRange =
         correctSourceRange(sourceRangeWithinAttribute, attributeOffset)!;
-    const sourceRangeLocationOffset: LocationOffset = {
-      line: sourceRange.start.line,
-      col: sourceRange.start.column,
-    };
-    const parseResult = parseJs(
-        expressionText,
-        attributeValueRange.file,
-        sourceRangeLocationOffset,
-        'polymer-expression-parse-error');
+    const parseResult = parseExpression(expressionText, sourceRange);
+    if (!parseResult) {
+      continue;
+    }
     if (parseResult.type === 'failure') {
       warnings.push(parseResult.warning);
     } else {
-      results.push(new ScannedDatabindingExpression(
+      results.push(new DatabindingExpression(
           node,
           attr,
           sourceRange,
@@ -307,4 +348,26 @@ function indexesToSourceRange(
     },
     end: {line: endLineNumLinesIntoText, column: endIndex - endOfLineIndex}
   };
+}
+
+function parseExpression(content: string, expressionSourceRange: SourceRange) {
+  const expressionOffset = {
+    line: expressionSourceRange.start.line,
+    col: expressionSourceRange.start.column
+  };
+  const parseResult = parseJs(
+      content,
+      expressionSourceRange.file,
+      expressionOffset,
+      'polymer-expression-parse-error');
+  if (parseResult.type === 'success') {
+    return parseResult;
+  }
+  // The polymer databinding expression language allows for foo.0 and foo.*
+  // formats when accessing sub properties. These aren't valid JS, but we don't
+  // want to warn for them either. So just return undefined for now.
+  if (/\.(\*|\d+)/.test(content)) {
+    return undefined;
+  }
+  return parseResult;
 }